move 'composeK' tests to separate file
