+++ types/ramda/tsconfig.json
@@ -44,6 +44,7 @@
         "test/clone-tests.ts",
         "test/comparator-tests.ts",
         "test/complement-tests.ts",
-        "test/compose-tests.ts"
+        "test/compose-tests.ts",
+        "test/composeK-tests.ts"
     ]
 }
